proyecto python/app/requirements/requirements.txt
asgiref==3.2.10
Django==3.1.1
django-crum==0.7.9
django-filter==2.4.0
django-widget-tweaks==1.4.8
freeze==3.0
Pillow==7.2.0
psycopg2==2.8.6
pytz==2020.1
six==1.15.0
sqlparse==0.3.1
